Markdown block: changes markdown parser
Replaces commonmark with Markdown-it. Commonmark was missing a key
feature: the parsed tree didn't have information of the delimiters used
by the user. This is crucial to the live preview feature: we need to
know if the user used `_` or `*` to be able to reproduce them in the
live-preview source.