Markdown block live preview: bug pasting markdown
When pasting Markdown code in the live-preview component, it was not
being rendered automatically. The user had to focus out of the live
preview component to make it work. This behaviour was caused by the
configuration passed to the observer method in the MutationObserver.
This observer listens for changes in the contenteditable element of the
Live Preview component. In order to be triggered by a paste event, the
`MutationObserverInit.childList` property should be passed as `true` to
the method configuration. But this triggers lots of node changes when
new code is pasted and it causes a significant performance problem, even
making the browser to freeze if the text is big enough.

Relying on the "onInput" handler is the best choice here, but
unfortunately, this doesn't work on Internet Explorer 11, so these users
will have to focus out of the Live Preview component in order to see
their Markdown code rendered.